barkat_wholesale/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.11.0
Django==5.2.8
django-cors-headers==4.9.0
django-tailwind==4.4.1
djangorestframework==3.16.1
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==12.0.0
pytailwindcss==0.3.0
python-bidi==0.6.7
python-dotenv==1.2.1
pytz==2025.2
pywin32==311
sqlparse==0.5.4
tzdata==2025.2
python-barcode==0.16.1
qrcode==8.0
pyzbar==0.1.9
